Add to Node an extension function, applying(_:range:), for applying a function to a particular range of a rope.  Use it to set, clear, and add attributes on a range of a rope.
While I'm here, make an extension public and the methods it adds
default visibility.

Add to the initializer init(content:) an attributes argument with
that defaults to the empty dictionary; now it is init(content:attributes:).

In Node.swift, import Foundation for NSAttributedString.Key.

Add to Node an extension method, apply<R>(_:at:base:), that applies
a function to the content at a particular UTF-16 offset in a rope.

Add to Node an extension property, length, that tells the Node's
length in UTF-16 units, and a method utf16(at:) that retrieves the
UTF-16 code unit at a particular offset.