Make Handle.Id an enum, .cursor(_), .index(_), .extent(_), to classify the kind of handle it is.  Each case has UInt64 associated data to identify individual handles.
Replace Set<Handle.Id> with HandleSet, which keeps count of the
number of handle identifiers of each kind inside in properties
cursorCount, extentCount, and indexCount.

Rename the Node.containsIndex() method to Node.containsHandle(),
since we really are testing for the presence of a handle, not the
index sub-kind of handle.