Add a routine for "tightening" a selection by stepping the left and right boundaries into the extent hierarchy as deep as possible without crossing over a character or an entire empty extent.
Add routines for "loosening" a selection on either the left or right
boundary by stepping the boundary out of the extent hierarchy as
high as possible without crossing over a character, an entire empty
extent, or the extent given in a `limit` argument.

Add auxiliary versions of index(before/after:climbing:) that tell
the deepest ExtentController at the given index, if any.

Add a helper routine, commonPrefix, that returns the longest common
prefix of two Sequences of Equatable.

Add tests for commonPrefix.

TBD tests for the other functions introduced by this change.

TBD improvements to < and == operators for RopeIndex.