Fix some bugs in Rope.Node.replacing() and extract some common code.
Finish adapting Rope.replace() to the new Rope.Node.replacing() API.

Update tests for the new .replacing() API that uses a `recording:
ChangeList` parameter instead of an `undoList: ChangeList`) .  All
RopeTests pass again.

Add a method for a ChangeList to append a second ChangeList to itself.

Add a `delegate: TypeErasedOffsetDelegate` property to Rope.  Make
a Rope its own default `delegate` with do-nothing implementations
of ropeDidChange(on:changeInLength:) and ropeAttributesDidChange(on:).

Add an extension method, indicateChanges<T>(new:old:undoList:), to
RopeOffsetDelegate.  The method indicates a rope change and records
the inverse rope change on the undo list.